Move CORS stuff to general route_helpers file
